Move models and controllers into separate directories
