+++ src/utils.js
@@ -1,0 +1,7 @@
+module.exports = {
+  generateKey: () => {
+    return [...Array(30)]
+      .map((e) => ((Math.random() * 36) | 0).toString(36))
+      .join('')
+  }
+}
